requirements.txt
google-genai
pyttsx3
SpeechRecognition
PyAudio
python-dotenv
